// client-app/src/models/Organization.model.ts
// Generated by https://quicktype.io

export interface Organization {
  id: number;
  createdAt: string;
  updatedAt: string;
  mediaUrl: null | string;
  name: string;
  email: string;
  password: string;
  phone: string;
  accessToken: null | string;
  description: null | string;
  webPage: null | string;
  instagramPage: null | string;
  facebookPage: null | string;
}

export interface OrganizationItem extends Organization {
  locations: Location[];
}

export interface Location {
  id: number;
  createdAt: string;
  updatedAt: string;
  organizationId: number;
  name: string;
  description: string;
  latitude: number;
  longitude: number;
  mediaUrls: string[];
}